Dynamically infer classpath by classloader according to the class usage (#14)
Dynamically infer classpath by classloader according to the class usage

When application is using play framework with hot-reload enabled, fast-avro
couldn't find the right classpath by extracting from system property:java.class.path.
So this code change will explicitly find all the classes being used in the
generated files and put the corresponding libraries in the compile classpath.
Play framework is using both current classloader and the thread context classloader
to discover class path in hot reload feature. In fast-avro class generation/compilation,
we have to do the same time to align with this behavior.
Once the generated file is compiled successfully, loading the compiled classes
is not a problem in play framework with hot-reload enabled.