Change to use IdentityHashMap to cache schemaId for Avro 1.4
In Avro 1.4, Schema#hashCode is not performance efficient. It has been improved
by caching computed hashcode in the later avro version.

Change to use IdentityHashMap for Avro 1.4 helps to improve the schemaId cache
lookup performance. Here's JMH benchmark of deserialization latency of a example
schema using different approaches:
 - vanilla avro 1.4:                                          0.26ms
 - avro-fastserde with ConcurrentHashMap as schemaId cache:   0.22ms
 - avro-fastserde with IdentityHashMap as schemaId cache:     0.15ms

However, it requires the client to resuse the schame instance. Or, it will eventually
lead to "Too many schema objects" issue. I think it's ok for now since a typical
client generally reuse the schema instance and that's one of the best practices as well.
We can build a LRU Cache with size limit in the future.