Fixed a bug in processUnion of FastSerializer
Previously, the code is checking schema type to decide the right index,
and this change was introduced to accommodate Avro-1.4, but this is
not appropriate since Union could contain multiple record types, which
will cause the writen index will always be the index of the first record.
The fix is to check the full name of the schema instead of type name.