Fix a bug in object reuse in CompositeByteBuffer (#48)
The buffer size should be less than or equal to the new size for object reuse to utilize the existing buffer in CompositeByteBuffer. The bug was that we previously did not check for equality, leading to spurious memory allocation in the common case where we attempt to reuse arrays with the same number of elements.