Made fast serde generation deterministic
This is achieved by altering the unique name generation strategy, such
that it is no longer globally-unique, but rather constrained to being
unique within a given serde class and furthermore within a given name
prefix. This should help remove as much needless jitter as possible in
the codegen output when the meta-code is altered.

Also created a new FastSerdeBase abstract class, extended by both the
FastDeserializerGeneratorBase and FastSerializerGeneratorBase. The
ser and deser meta-code is pretty repetitive, so I'm hoping that we
can move more code to this shared ancestor class later down the line.