Refined union tests.
Added a scenario to the union of records where multiple records with
different names can be present, in different orders.

Added another scenario where the reader schema could be missing some
union branches, and Avro is expected to fail at run-time if the missing
branch is exercised in the written data (but will succeed otherwise).

Code-gen for the tests is included in this commit to act as a baseline
for subsequent commits.