Enhanced fast-avro tests to leverage reflection for SpecificRecords
This is in preparation of support Avro 1.9 and 1.10 which do not
generate public fields by default. The new reflection-based method
will look for either the public field or the getter, which makes
the tests work for both styles of SpecificRecord, and thus work
for all Avro versions between 1.4 and 1.10.

Also added two new Utils functions to more accurately determine if
the Avro version on the classpath supports non-Utf8 strings. The
tests previously assumed that only 1.4 lacked support for non-Utf8
strings, but actually the support was added gradually throughout
1.6 and 1.7.

Finally, changed the stringableTest.avsc schema so that it adds
one more type in the union of the stringable field, as that is
necessary to sidestep a bug in 1.9 where the code-generated SR
produced by Avro does not even compile... this is later fixed by
1.10, but it makes 1.9 testing support impossible, so we must
sidestep the issue in this way.