Replaced 'JSwitch' usage by `JConditional` because of a JDK bug
Switch usage in the generated fast classes could potentially trigger
this jdk bug: https://bugs.java.com/bugdatabase/view_bug.do?bug_id=8252029
Which would crash the application if the application is using a lower
version than 11.0.11.
There is no way to force every user to use the fixed version because of
various reasons, so we will temporarily replace all the `switch` usage
by `if/else`, which may cause some performance regression if there
are too many branches, but for the majority of the use cases, the performance
should be similar.
The places updated in this PR is the deserializer functions for union and enum.
For performance reason, this PR builds a lookup for enum processing, hope
the performance is good enough even without switch.