[helper] Fix broken escaping in codegen by Avro 1.4/1.5. (#254)
Avro 1.4/1.5 generate broken (uncompilable) code when the schema has
escaped chars in strings. We fix such code in the compat helper by:
* First, undoing the damage (unescaping `\"` back to `"`),
* Then, escaping backslashes (escaping `\` to `\\`), and
* Finally, re-escaping the quotes (from `"` to `\"`).

If we are replacing the schema with an "alternative", since the alt
comes directly from JSON, and not from codegen, we escape it using
a standard function (`StringEscapeUtils.escapeJava`), and don't bother
fixing up the original broken code.

If the code was generated by Avro 1.6 or later, it would be correctly
escaped to begin with, so there's nothing more to do.

The test was copied from PR #252, courtesy of @radai-rosenblatt.

Fixes #253.