Tweaked the fast-avro schema fingerprinting logic
It used to leverage the Avro fingerprinting, which is based on the
"parsing canonical form" of the schema, but this ignores properties
such as which class to use for String deserialization. This is a
problem in cases where we wish to deserialize the same schema both
with and without Java Strings, since the cache will return the same
deserializer for both, ultimately leading to class cast issues.

The new approach is to simply take the hash code of the full string
of the schema.