Fix the map/array for nested maps/array with unions at play (#578)
Missed a few cases in #576 when we have nested maps or array with the following cases:

In case we have a map<map> inside a union, the old logic was passing a production which had the map-start,map-end symbols. However while processing maps our new map logic expects to see a map-end symbol as the action which we then check the production for to find the actions symbols for the map value. In this PR, we change union processing to seek to this action before passing to the map. Once again this indicates that we need a rewrite of this entire action processing logic to standardize which symbols are expected when we process any type to avoid unnecessary seeks and reduce confusion.
If we have map<map> on writer and map<union<map> on reader there's actually no ambiguity about which map-end symbol to use for the value action since the inner map symbol is now contained in a UnionAdjustAction. We account for this by modifying our checks in processMap and processArray